Release v4.5

+++ arch/platform/jn516x/dev/jn516x-ccm-star.c
@@ -48,8 +48,8 @@ static int current_key_is_new = 1;
 /*---------------------------------------------------------------------------*/
 static void
 aead(const uint8_t *nonce,
-     uint8_t *m, uint8_t m_len,
-     const uint8_t *a, uint8_t a_len,
+     uint8_t *m, uint16_t m_len,
+     const uint8_t *a, uint16_t a_len,
      uint8_t *result, uint8_t mic_len,
      int forward)
 {
