+++ os/lib/ccm-star.h
@@ -54,25 +54,25 @@
  * Structure of CCM* drivers.
  */
 struct ccm_star_driver {
-  
+
   /**
    * \brief         Sets the key in use. Default implementation calls AES_128.set_key().
    * \param key     The key to use.
    */
   void (* set_key)(const uint8_t* key);
-  
+
   /**
    * \brief         Combines authentication and encryption.
    * \param nonce   The nonce to use. CCM_STAR_NONCE_LENGTH bytes long.
-   * \param m       message to encrypt or decrypt
-   * \param a       Additional authenticated data
+   * \param m       message to encrypt or decrypt. Up to 0xffff
+   * \param a       Additional authenticated data. Up to 0xfeff
    * \param result  The generated MIC will be put here
    * \param mic_len The size of the MIC to be generated. <= 16.
    * \param forward != 0 if used in forward direction.
    */
   void (* aead)(const uint8_t* nonce,
-      uint8_t* m, uint8_t m_len,
-      const uint8_t* a, uint8_t a_len,
+      uint8_t* m, uint16_t m_len,
+      const uint8_t* a, uint16_t a_len,
       uint8_t *result, uint8_t mic_len,
       int forward);
 };
